some code review cleanup
