+++ src/ent/small_move_arena.h
@@ -1,6 +1,7 @@
 #ifndef ARENA_H
 #define ARENA_H
 
+#include "../compat/malloc.h"
 #include "../util/log.h"
 
 #include <assert.h>
@@ -15,43 +16,6 @@ typedef struct {
 } Arena;
 
 #define INITIAL_ARENA_CAPACITY (1024 * 1024) // 1 MB
-
-#ifdef _WIN32
-#include <malloc.h> // For _aligned_malloc and _aligned_free
-#endif
-
-// Portable aligned allocation function
-static inline int portable_aligned_alloc(void **ptr, size_t alignment,
-                                         size_t size) {
-  if (!ptr) {
-    return -1; // Invalid pointer
-  }
-
-#ifdef _WIN32
-  // Windows-specific aligned allocation
-  *ptr = _aligned_malloc(size, alignment);
-  if (*ptr == NULL) {
-    return -1; // Allocation failed
-  }
-  return 0; // Success
-#else
-  // POSIX-compliant aligned allocation
-  // posix_memalign returns 0 on success, or an error number on failure
-  return posix_memalign(ptr, alignment, size);
-#endif
-}
-
-// Portable aligned free function
-static inline void portable_aligned_free(void *ptr) {
-  if (!ptr)
-    return;
-
-#ifdef _WIN32
-  _aligned_free(ptr);
-#else
-  free(ptr);
-#endif
-}
 
 static inline Arena *create_arena(size_t initial_capacity, size_t alignment) {
   // alignment = 16 for our "SmallMove" structure.
@@ -80,7 +44,9 @@ static inline void *arena_alloc(Arena *arena, size_t size) {
   }
 
   // Ensure 'size' is a multiple of 16 for 16-byte alignment
-  assert(size % 16 == 0 && "Allocation size must be a multiple of 16 bytes.");
+  if (size % 16 != 0) {
+    log_fatal("Allocation size must be a multiple of 16 bytes.");
+  }
 
   // Check if there's enough space; if not, grow the arena
   if (arena->size + size > arena->capacity) {
@@ -118,7 +84,9 @@ static inline void arena_dealloc(Arena *arena, size_t size) {
   }
 
   // Ensure 'size' is a multiple of 16 for 16-byte alignment
-  assert(size % 16 == 0 && "Allocation size must be a multiple of 16 bytes.");
+  if (size % 16 != 0) {
+    log_fatal("Allocation size must be a multiple of 16 bytes.");
+  }
   arena->size -= size;
 }
